fixed win32 compilation
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@2815 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ src/player_manager.cpp
@@ -425,7 +425,7 @@ void IPlayerManager::updatePlayers() {
 					const v2<int> &checkpoint_size = _checkpoints[c].size;
 					const v3<int> &checkpoint_pos  = _checkpoints[c].position;
 					{
-						int yn = (int) sqrt(checkpoint_size.y * _players.size() / checkpoint_size.x);
+						int yn = (int) sqrt((double)(checkpoint_size.y * _players.size() / checkpoint_size.x));
 						int xn = (_players.size() - 1) / yn + 1;
 						int n = xn * yn;
